refactor Cobb County core out
